Added contact, got rid of admin
Removed public link to admin dashbaord, added a new page called contact and moved things from the about into it

Patched some bugs as well, commented out the sms so as to not go over our twilio budget
