+++ react-app/src/css/Style.css
@@ -48,6 +48,12 @@ p {
     accent-color: #2c4356;
 }
 
+.step-header {
+    padding: 2.5vh;
+    font-family: "Bearhug";
+    font-size: 5vh;
+}
+
 .loading {
     margin-left: auto;
     margin-right: auto;
@@ -58,6 +64,7 @@ p {
 
 .radio-input {
     width: auto;
+    padding: 2.5vh
 }
 
 label {
